Moved crazy CBiRRT serialization into CBiRRT.
Nothing-- I repeat NOTHING-- other than CBiRRT should use this
serialization of TSRs.  Everything should use JSON, YAML, or a dict.
